Move Trackers to namespace to hide them
