+++ analyzers/src/SonarAnalyzer.CSharp/Trackers/CSharpBaseTypeTracker.cs
@@ -24,7 +24,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Helpers.Trackers
 {
     public class CSharpBaseTypeTracker : BaseTypeTracker<SyntaxKind>
     {
