+++ analyzers/src/SonarAnalyzer.CSharp/Trackers/CSharpElementAccessTracker.cs
@@ -23,7 +23,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using SonarAnalyzer.Extensions;
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Helpers.Trackers
 {
     public class CSharpElementAccessTracker : ElementAccessTracker<SyntaxKind>
     {
