+++ analyzers/src/SonarAnalyzer.CSharp/Trackers/CSharpFieldAccessTracker.cs
@@ -23,7 +23,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Helpers.Trackers
 {
     public class CSharpFieldAccessTracker : FieldAccessTracker<SyntaxKind>
     {
