+++ analyzers/src/SonarAnalyzer.CSharp/Trackers/CSharpObjectInitializationTracker.cs
@@ -28,7 +28,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using SonarAnalyzer.Extensions;
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Helpers.Trackers
 {
     /// <summary>
     /// Verifies the initialization of an object, whether one or more properties have been correctly set when the object was initialized.
