+++ analyzers/src/SonarAnalyzer.Common/Facade/ITrackerFacade.cs
@@ -18,6 +18,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using SonarAnalyzer.Helpers.Trackers;
+
 namespace SonarAnalyzer.Helpers.Facade
 {
     public interface ITrackerFacade<TSyntaxKind>
