+++ analyzers/src/SonarAnalyzer.VisualBasic/Trackers/VisualBasicBaseTypeTracker.cs
@@ -24,7 +24,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.VisualBasic;
 using Microsoft.CodeAnalysis.VisualBasic.Syntax;
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Helpers.Trackers
 {
     public class VisualBasicBaseTypeTracker : BaseTypeTracker<SyntaxKind>
     {
