+++ analyzers/src/SonarAnalyzer.VisualBasic/Trackers/VisualBasicElementAccessTracker.cs
@@ -22,7 +22,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.VisualBasic;
 using Microsoft.CodeAnalysis.VisualBasic.Syntax;
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Helpers.Trackers
 {
     public class VisualBasicElementAccessTracker : ElementAccessTracker<SyntaxKind>
     {
