+++ analyzers/src/SonarAnalyzer.VisualBasic/Trackers/VisualBasicFieldAccessTracker.cs
@@ -22,9 +22,8 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.VisualBasic;
 using Microsoft.CodeAnalysis.VisualBasic.Syntax;
-using SonarAnalyzer.Common;
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Helpers.Trackers
 {
     public class VisualBasicFieldAccessTracker : FieldAccessTracker<SyntaxKind>
     {
